Rename file name and add other test cases.
